create cache key builder
