+++ src/Kros.AspNetCore/Authorization/CachedJwtTokenProvider.cs
@@ -4,7 +4,6 @@ using Microsoft.Extensions.Options;
 using Microsoft.Extensions.Primitives;
 using System;
 using System.Linq;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
 namespace Kros.AspNetCore.Authorization;
@@ -14,13 +13,11 @@ namespace Kros.AspNetCore.Authorization;
 /// </summary>
 internal class CachedJwtTokenProvider : IJwtTokenProvider
 {
-    private const string CacheKeyPrefix = "JwtToken:";
-
     private readonly ICacheService _cacheService;
     private readonly IHttpContextAccessor _httpContextAccessor;
     private readonly GatewayJwtAuthorizationOptions _jwtAuthorizationOptions;
     private readonly IJwtTokenProvider _jwtProvider;
-    private static Regex _cacheRegex = null;
+    private readonly ICacheKeyBuilder _cacheKeyBuilder;
 
     /// <summary>
     /// Initializes a new instance of the CachedJwtTokenProvider class.
@@ -29,21 +26,19 @@ internal class CachedJwtTokenProvider : IJwtTokenProvider
     /// <param name="httpContextAccessor">The HTTP context accessor.</param>
     /// <param name="jwtAuthorizationOptions">Authorization options.</param>
     /// <param name="jwtProvider">The JWT provider instance.</param>
+    /// <param name="cacheKeyBuilder">The cache key builder instance.</param>
     public CachedJwtTokenProvider(
         ICacheService cacheService,
         IHttpContextAccessor httpContextAccessor,
         IOptions<GatewayJwtAuthorizationOptions> jwtAuthorizationOptions,
-        IJwtTokenProvider jwtProvider)
+        IJwtTokenProvider jwtProvider,
+        ICacheKeyBuilder cacheKeyBuilder)
     {
         _cacheService = Check.NotNull(cacheService, nameof(cacheService));
         _httpContextAccessor = Check.NotNull(httpContextAccessor, nameof(httpContextAccessor));
         _jwtAuthorizationOptions = Check.NotNull(jwtAuthorizationOptions.Value, nameof(jwtAuthorizationOptions));
         _jwtProvider = Check.NotNull(jwtProvider, nameof(jwtProvider));
-
-        if (!string.IsNullOrWhiteSpace(_jwtAuthorizationOptions.CacheKeyUrlPathRegexPattern))
-        {
-            _cacheRegex = new Regex(_jwtAuthorizationOptions.CacheKeyUrlPathRegexPattern);
-        }
+        _cacheKeyBuilder = Check.NotNull(cacheKeyBuilder, nameof(cacheKeyBuilder));
     }
 
     /// <inheritdoc/>
@@ -56,7 +51,7 @@ internal class CachedJwtTokenProvider : IJwtTokenProvider
             return await _jwtProvider.GetJwtTokenAsync(token);
         }
 
-        string cacheKey = BuildCacheKey(_httpContextAccessor.HttpContext, token);
+        string cacheKey = _cacheKeyBuilder.BuildCacheKey(_httpContextAccessor.HttpContext, token);
 
         return await _cacheService.GetOrCreateAsync(
             cacheKey,
@@ -73,41 +68,12 @@ internal class CachedJwtTokenProvider : IJwtTokenProvider
             return await _jwtProvider.GetJwtTokenForHashAsync(hashValue);
         }
 
-        string cacheKey = BuildHashCacheKey(hashValue);
+        string cacheKey = _cacheKeyBuilder.BuildHashCacheKey(hashValue);
 
         return await _cacheService.GetOrCreateAsync(
             cacheKey,
             () => _jwtProvider.GetJwtTokenForHashAsync(hashValue));
     }
-
-    /// <summary>
-    /// Builds a cache key based on the HTTP context and token.
-    /// </summary>
-    /// <param name="httpContext">The HTTP context.</param>
-    /// <param name="token">The authorization token.</param>
-    /// <returns>The cache key.</returns>
-    private string BuildCacheKey(HttpContext httpContext, StringValues token)
-    {
-        CacheHttpHeadersHelper.TryGetValue(
-            httpContext.Request.Headers,
-            _jwtAuthorizationOptions.CacheKeyHttpHeaders,
-            out string cacheKeyPart);
-
-        string urlPathForCache = GetUrlPathForCacheKey(httpContext);
-        if (urlPathForCache != null)
-        {
-            cacheKeyPart += urlPathForCache;
-        }
-
-        return GetKey(token, cacheKeyPart);
-    }
-
-    /// <summary>
-    /// Builds a cache key for hash-based authorization.
-    /// </summary>
-    /// <param name="hashValue">The hash value.</param>
-    /// <returns>The cache key.</returns>
-    private static string BuildHashCacheKey(StringValues hashValue) => GetKey(hashValue.ToString());
 
     /// <summary>
     /// Determines whether caching should be used for the given request.
@@ -135,36 +101,4 @@ internal class CachedJwtTokenProvider : IJwtTokenProvider
     private bool IsRequestPathIgnoredForCache(HttpRequest request)
         => _jwtAuthorizationOptions.IgnoredPathForCache
         .Contains(request.Path.Value.TrimEnd('/'), StringComparer.OrdinalIgnoreCase);
-
-    /// <summary>
-    /// Gets the URL path for cache key based on regex pattern.
-    /// </summary>
-    /// <param name="httpContext">The HTTP context.</param>
-    /// <returns>The URL path for cache key or null if not found.</returns>
-    private string GetUrlPathForCacheKey(HttpContext httpContext)
-    {
-        if (!string.IsNullOrWhiteSpace(_jwtAuthorizationOptions.CacheKeyUrlPathRegexPattern)
-            && !string.IsNullOrWhiteSpace(httpContext.Request.Path)
-            && _cacheRegex != null)
-        {
-            var match = _cacheRegex.Match(httpContext.Request.Path);
-            if (match.Success)
-            {
-                return match.Groups.Values.Last().Value;
-            }
-        }
-        return null;
-    }
-
-    /// <summary>
-    /// Generates a hash key from the given values.
-    /// </summary>
-    /// <param name="value">The primary value.</param>
-    /// <param name="additionalKeyPart">Additional key part (optional).</param>
-    /// <returns>The hash code.</returns>
-    private static string GetKey(StringValues value, string additionalKeyPart = null)
-    {
-        int key = (additionalKeyPart is null) ? HashCode.Combine(value) : HashCode.Combine(value, additionalKeyPart);
-        return CacheKeyPrefix + key.ToString();
-    }
 }
